Move ingest-event file
