+++ src/worker/worker.ts
@@ -1,4 +1,3 @@
-import { ingestEvent } from '../ingestion/ingest-event'
 import { initApp } from '../init'
 import { runPlugins, runPluginsOnBatch, runPluginTask } from '../plugins/run'
 import { loadSchedule, setupPlugins } from '../plugins/setup'
@@ -6,6 +5,7 @@ import { createServer } from '../server'
 import { status } from '../status'
 import { PluginsServer, PluginsServerConfig } from '../types'
 import { cloneObject } from '../utils'
+import { ingestEvent } from './ingestion/ingest-event'
 
 type TaskWorker = ({ task, args }: { task: string; args: any }) => Promise<any>
