+++ tests/postgres/worker.test.ts
@@ -3,13 +3,13 @@ import IORedis from 'ioredis'
 import { mocked } from 'ts-jest/utils'
 
 import Client from '../../src/celery/client'
-import { ingestEvent } from '../../src/ingestion/ingest-event'
 import { loadPluginSchedule } from '../../src/main/services/schedule'
 import { runPlugins, runPluginsOnBatch, runPluginTask } from '../../src/plugins/run'
 import { loadSchedule, setupPlugins } from '../../src/plugins/setup'
 import { ServerInstance, startPluginsServer } from '../../src/server'
 import { LogLevel } from '../../src/types'
 import { delay, UUIDT } from '../../src/utils'
+import { ingestEvent } from '../../src/worker/ingestion/ingest-event'
 import { makePiscina } from '../../src/worker/piscina'
 import { createTaskRunner } from '../../src/worker/worker'
 import { resetTestDatabase } from '../helpers/sql'
@@ -17,7 +17,7 @@ import { setupPiscina } from '../helpers/worker'
 
 jest.mock('../../src/sql')
 jest.mock('../../src/status')
-jest.mock('../../src/ingestion/ingest-event')
+jest.mock('../../src/worker/ingestion/ingest-event')
 jest.mock('../../src/plugins/run')
 jest.mock('../../src/plugins/setup')
 jest.setTimeout(600000) // 600 sec timeout
